mapreduce: extract BackgroundStopper into its own module
Move the workaround class to background_stopper.py with a top-of-file
note that it exists as a workaround for an mngr-core issue and can be
deleted (along with its import in orchestration.py) once the core
Host.stop_agents enforces its own wall-clock budget.
